+++ R/format_label.R
@@ -1,0 +1,9 @@
+format_label <- function(x) {
+
+  x %>%
+    stringr::str_replace_all(
+      "([#$%&_{}])",
+      "\\\\\\1"
+    )
+
+}